LPS-61952 Move calendar exceptions
